zuri commit

+++ index.js
@@ -2,8 +2,8 @@ let operand1 = prompt("Enter first number: ");
 let operators = prompt("Enter operator: +, -, *, /, **, %");
 let operand2 = prompt("Enter second number: ");
 
-operand1 = parseInt(operand1);
-operand2 =parseInt(operand2)
+operand1 = parseFloat(operand1);
+operand2 = parseFloat(operand2)
 
 let result = 0;
 // let error = "No arithmetic operator defined!";
